azure_function_cluster/ClusterIncidents/requirements.txt
pandas
numpy
hdbscan
azure-functions

